+++ gfa-anagrafiche/src/main/java/gov/mef/gfa/anagrafiche/dao/EnteDao.java
@@ -1,5 +1,7 @@
 package gov.mef.gfa.anagrafiche.dao;
 
+import java.math.BigDecimal;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -7,5 +9,5 @@ import gov.mef.gfa.anagrafiche.model.Ente;
 
 @Repository
 public interface EnteDao extends JpaRepository<Ente, Long> {
-	public Ente findByIdente(Integer id);
+	public Ente findByIdente(BigDecimal id);
 }
