+++ gfa-anagrafiche/src/main/java/gov/mef/gfa/anagrafiche/service/EnteService.java
@@ -1,8 +1,10 @@
 package gov.mef.gfa.anagrafiche.service;
 
+import java.math.BigDecimal;
+
 import gov.mef.gfa.anagrafiche.bean.EnteRes;
 import gov.mef.gfa.anagrafiche.excepltion.ServiceException;
 
 public interface EnteService {
-	public EnteRes getEnteById(Integer id) throws ServiceException;
+	public EnteRes getEnteById(BigDecimal id) throws ServiceException;
 }
